+++ pkg/runtime/vmlimits_openbsd.go
@@ -13,7 +13,7 @@
 
 // +build openbsd
 
-package main
+package runtime
 
 import (
 	"syscall"
